Change drag image
The custom drag image stopped working on Chrome and Safari,
this is because the DOM element needs to be 'in view', and it
was off screen at the time of the drag.

This change builds the DOM element at the time of the drag
and places it at the top-level, last item in the body, but
slightly offscreen (above the header). It then removes the
element on dragEnd.